예외 처리, url 정리

See merge request s08-final/S08P31B208!31

+++ back/dubeng-dublist/src/main/java/com/ssafy/dubengdublist/repository/DubKingRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface DubKingRepository extends JpaRepository<DubKing, Integer>, DubKingRepositoryCustom {
+public interface DubKingRepository extends JpaRepository<DubKing, Long>, DubKingRepositoryCustom {
